A new low-level text management library (canvas for ultrafast compositing of unicode letters) is now separated from the higher level rendering and I/O (that is, libcaca). This commit totally breaks the API, but once everything is polished I will think about source-level backward compatibility. Most drivers are broken, but X11 still sorta works.
  The new design is much more object-oriented and allows having several
active renderers at the same time, changing renderers on the fly, and more
important, having no renderer at all (useful for converters, or when you
want to do your own renderer).

  And in case you are still wondering, the libcucul acronym has "Unicode"
because I want to support at least a subset of Unicode. There are awesome
glyphs in it, including the ones inherited from Codepage 437 such as
"gray 25%" that are used in DOS and Win32 ANSI art.