Revert setlocale() effect as soon as possible. Closes #12.
The Xlib and ncurses libraries query the LC_CTYPE locale value to guess
the usable character set. If the calling program did not call setlocale(),
that character set will be severely limited. Extensive Unicode support
is a reasonable libcaca user expectation.

The locale is restored as soon as possible, once the window or terminal
have been initialised. Unfortunately, the effect of setlocale() is process-
wide, and may affect other threads. This is now documented.

Note also that both Xlib and ncurses ignore the effects of uselocale()
which would have been the thread-safe solution to this problem.